refactor: int -> BigInteger 타입변경

+++ src/main/java/calculator/Application.java
@@ -1,9 +1,13 @@
 package calculator;
 
+import java.math.BigInteger;
+
 public class Application {
     public static void main(String[] args) {
         String input = InputHandler.readInput();
-        int result = Calculator.calculate(Parser.parse(input));
+
+        BigInteger result = Calculator.calculate(Parser.parse(input));
+
         OutputHandler.printOutput(result);
     }
 }
